Changes how precision of returned value is handled
I removed the precision parameter from the 'convert' function and I am now using 'decimal.Decimal' to handle setting to float precision. The trailing zeros of a float in `decimal.decimal` are preserved and is what I am using to set the precision. This keeps with the SI standard. If someone passes an integer `decimal.Decimal` instance the returned value would be an integer. if the passed value is an integer the returned value  is an integer as well. If the passed value is a float the return value is float with maximum precision and if passing a float `decimal.Decimal` instance the returned value is a float rounded to the precision gotten from the passed value.