Rewrite to add  support for more units
This rewrite adds mechanics that allows the addition of units either by the library maintainer or another person. It also allows the addition of units in user code if needed.

I have begun the bug testing of the code and there are still problems that need to be ironed out. It is working but there is an issue with some of the conversions that I have to sort out;