Several changes
Moves temperature conversions into the Unit class. The temperature conversions will not function just like the rest of the conversions.

Added a little more information to the documentation.

Added ft_survey unit.

Increased the precision on several units.

Documentation is going to be a long task that I could use soe help with. I need to separate the units into groups and I also need to provide full names and what exactly they are for. At the same tie if the precision of the unit can be improved then it should be done at that time. It is quite a bit of scowering the internet to find the best precision. At the same time if we happen to come across other units to add we should do so at that time.