Splits unit converter into several files
The code base was getting to long for a single file so for the sake of organization I split it into several files.

The other thing that splitting it into several files allowed me to do was put into place some voodoo code for the unit constants. so now instead of unit_converter.Unit.m it's unit_converter.units.m. Here is the voodoo code bit. To define constants for all of the units is one thing but adding constants for all of the units with all of the prefixes would have been insane. So what the voodoo magic is you can now do the following, unit_converter.units.cm or unit_converter.units.mm and a unit for 'mm' will be created on demand.

I also added support for ^ followed by a numerical value that can be used to signify exponents when passing the unit as a string to `convert` or passing the unit to Unit and not using the exponent keyword.