Changes returned value type
`int` return type has been removed. a `float` will always be returned.

if an `int` is passed into the converter it will be treated the same as a `float` being passed in and the returned value will have maximum precision.

I also added passing a `str` or `bytes` and these are treated the same as passing a `decimal.Decimal` instance. trailing zeros  of a fractional `decimal.Decimal`, `str` or `bytes` instance are treated as precision.