Adds quantities and constants
OK I added the SI constants so the accuracy will be improved. I still have more work to do with it like wrapping all of the numbers in `Decimal`. It works except for floating point math errors which will be taken care of once I wrap the numbers.

This next bit is going to be a tad long because I have to explain it from the bottom up so it will make sense.

All units in the SI system break down into 1 or a combination of 7 core units, This is called the "dimension". What the dimension is used for is to know what units can be converted between each other. I understand the purpose for it but it is really not needed in order to know what units are able to be converted between each other. Let me explain.

There are what is called quantities in the SI system. Now why on earth they called it by that name is beyond me and as much as I search the internet I am not able to find a list of all of the quantities the SI system has. I can find lists for specific fields of study like mechanical and electrical. The issue with trying to put together a list of the quantities in this manner is I do not know all of the fields of study that use the SI system and I cannot find a list of all of the fields of study that use the SI system. Very annoying.

Quantities are simply groups of units that can be converted between each other. Each group has a unit as it's base unit. This is what all units in that group will convert to. This provides the ability to supply aa single factor and be able to convert between any unit in the group. If a unit is a "from" unit it divides the input value by the factor to achieve the group unit and a "to" unit will multiply by the factor to convert from the group unit.

The quantity will typically have it's own symbol just like a unit does, I am not using this symbol at all I just thought it was worth mentioning.

Because the quantity has a specific unit it uses it is not hard to locate either the base SI symbol that represents the unit or it will be a symbol pointing to units outside of the quantity group and from those you can get the SI representations of.

As long as 2 units have the same base SI symbol for the unit that is the "factor" or go between unit for the conversion the units are compatible. That being said you can convert between units that are not in the same quantity group so long as that SI symbol matches.

So I added 169 SI categories along with the SI symbol for the categories unit and I also converted those SI symbols into a base SI symbol and that is when the magic really starts to happen. Once everything is broken down into the base SI symbol units and quantities alike I am now able to connect a unit units that are compatible for conversion. I am also able to link a unit to compatible quantities as well.

Now I am nto finished with the modifications just yet as I still have code cleanup to do (large change) and I still have some ideas rolling around in my head for some other things to add.

I did want to mention that this addition is not going to list units tha are dynamically made or that are user input and there is no pre defined unit for it. an example is miles per hour. While I did hard code in the ability to use "mph" this is an incorrect symbol for miles per hour. The correct symbol is "mi/h" or "mi h^-1".  so if a user selects the unit "mph" and prints out a list of compatible units "mi/h" is ot going to show up in the list. That is because "mi/h" is not a single unit. It is 2 units written as a mathematical equation. If I was to back pedal starting form the 7 core SI units and located every possible combination of units that can mathematically be put together and would be valid for a unit conversion the list would be HUGE and to generate that list would take a large amount of time. That list would also have in it some off the wall combinations of units that really do not relate to each other.

I did also want to mention that I found a couple of bugs during this process and I have fixed them. One of those bugs caused me to write a new unit parser. This new parser is faster and easier to understand what is happening.

Look at the example file at the bottom it shows how to use the new features. I have not yet documented the features  as I am still getting a better understanding of the relationships between quantities and units.