New quantity system is done.
This is another big update to the program.

The quantity system is designed in a manner that allows a user to add additional quantities and that gets done by subclassing the `Quantities` class. There is a metaclass attached to `Quantities` that handles injecting the newly created class. An instance of the class does not need to be made.

It is important to use CaMeL case for the class name. The class name gets parsed and a space is added before every uppercase letter (except  the first) and that gets set to the class attribute `name`. `name` does not have to be set manually unless you want the name to to be different then the class name.

The docstring gets turned into a single line description that is accessible from the description class attribute.

You do not need to set the description attribute unless you want the description to be different then the docstring.

The symbol attribute is optional. It is a nicety that just gives more information is wanted.  At this time it does not effect hot the program runs at all.

The unit attribute is mandatory, this is what the program uses to determine what units are compatible. Each quantity in the SI system typically has a unit attached to it. This unit is what units gets converted to and converted from. It is what all conversion factors for a specific quantity are set for.

How the program check for compatible units is every unit in the SI system breaks down into one or more of 7 base units. by comparing the "SI equivalent" or "SI expression" for a quantity to the same for another quantity if they match then the quantities are compatible and conversions can be done across them.

Here is an example of how to correctly add a quantity to the program

```python

from python_utils.unit_converter.quantity import Quantities

class Crackle(Quantities):
    """
    Change of jounce per unit time: the fifth time derivative of position
    """
    symbol = 'c'
    unit = units.m / units.s(exponent=5)
```

The quantities in the SI system are wild as there are only a small number of them that are in the SI standard. Most of the SI quantities are created based on use of the SI system. Here are some examples.

Physics
Chemistry
Mechanical
Radiometry
Optics
Molar
Electromagnetic
Photometric

Each of those fields has SI quantities that relate to different unit groupings. I want to note that there are overlaps between the different fields and also with the quantities written into the SI standard. They may share the same name and they may not. If the name is not shared but you would like to have the correct name attached to the quantity then subclass the quantity you want  to create a new name for. You do not have to specify the unit when doing this if the unit is the same. All other attributes function the same way.

Now remember an instance of the new quantity does not need to be created for this to work    you simply hve to define the quantity class and either subclass `Quantities` or another quantity class.

I have done a large amount of trial and error in order to get the best possible speed while keeping the memory footprint as small as possible. In the  example file at the end I have the system enumerate all of the core units by way of iterating over all available quantities and the compatible units for those quantities.

There are 514 core units and 179 core quantities I have added so far. I still have to add the units and quantities related to Optics and any other fields of study that use the SI system. I have not been able to locate a list fo field of study that use the SI system so locating them has been a matter of stumbling across them on accident.