Adds 2 way conversions
This is kinda cool.

```python
mph = units.mph
kph = units.kph
mph /= kph

mph(20)
print(kph.value)

kph(150)
print(mph.value)
```

so when setting the value of one unit the cnoverted value shows up on the other.  Neat thing about that isit allows 2 remote sections of code in a program to work without actually passing anything between them. An example is say you have one block of code that hands the input and output of data. and that data is in kph, then you have another bunch of code that handles say a GUI interface. when the program initially starts a module level attribute can be created for the conversion using this technique. The data in and out uses the "kph" attribute and the gui uses the "mph" attribute. so the gui would do it's thing   and when updating the information it would  collect the value of "mph" the data end would set the data into "kph" which does the conversion and spits it out on the other end in the "mph" attribute.

It's kinda cool actually.

I can even write it so it supports more then a single conection. so if you wanted to say convert from km/h to mi/h, ft/s and m/s that could be done pretty easily. as it stands right now this can be done.

```python
kph = units.kph
kph.value = 60

kph.mph
kph.fps
kph.mps
```
and the returned value would be the converted value.